+++ tests/Nexendrie/Rss/CloudTest.php
@@ -15,14 +15,14 @@ final class CloudTest extends \Tester\TestCase
 {
     public function testPort(): void
     {
-        $cloud = new Cloud("test.com", 80, "/test", "test.a", "http-post");
+        $cloud = new Cloud("test.com", 80, "/test", "test.a", CloudProtocol::HttpPost);
         $cloud->port = 1;
         Assert::same(1, $cloud->port);
     }
 
     public function testPath(): void
     {
-        $cloud = new Cloud("test.com", 80, "/test", "test.a", "http-post");
+        $cloud = new Cloud("test.com", 80, "/test", "test.a", CloudProtocol::HttpPost);
         $cloud->path = "/abc";
         Assert::same("/abc", $cloud->path);
         Assert::exception(function () use ($cloud) {
@@ -30,26 +30,16 @@ final class CloudTest extends \Tester\TestCase
         }, \InvalidArgumentException::class);
     }
 
-    public function testRegisterProtocol(): void
-    {
-        $cloud = new Cloud("test.com", 80, "/test", "test.a", "http-post");
-        $cloud->protocol = "soap";
-        Assert::same("soap", $cloud->protocol);
-        Assert::exception(function () use ($cloud) {
-            $cloud->protocol = "abc";
-        }, \InvalidArgumentException::class);
-    }
-
     public function testAppendToXml(): void
     {
-        $cloud = new Cloud("test.com", 80, "/test", "test.a", "http-post");
+        $cloud = new Cloud("test.com", 80, "/test", "test.a", CloudProtocol::HttpPost);
         $xml = new \SimpleXMLElement("<test></test>");
         $cloud->appendToXml($xml);
         Assert::same($cloud->domain, (string) $xml->cloud["domain"]);
         Assert::same((string) $cloud->port, (string) $xml->cloud["port"]);
         Assert::same($cloud->path, (string) $xml->cloud["path"]);
         Assert::same($cloud->registerProcedure, (string) $xml->cloud["registerProcedure"]);
-        Assert::same($cloud->protocol, (string) $xml->cloud["protocol"]);
+        Assert::same($cloud->protocol->value, (string) $xml->cloud["protocol"]);
     }
 }